⚡️ (optimize-packing) Use uint96 to pack Account tightly

+++ test-foundry/TestHeapOrdering.t.sol
@@ -21,8 +21,8 @@ contract TestHeapOrdering is DSTest {
 
     function update(
         address _id,
-        uint256 _formerValue,
-        uint256 _newValue
+        uint96 _formerValue,
+        uint96 _newValue
     ) public {
         heap.update(_id, _formerValue, _newValue, MAX_SORTED_USERS);
     }
@@ -188,7 +188,7 @@ contract TestHeapOrdering is DSTest {
 
     function testShouldInsertAccountsAllSorted() public {
         for (uint256 i = 0; i < accounts.length; i++) {
-            update(accounts[i], 0, NB_ACCOUNTS - i);
+            update(accounts[i], 0, uint96(NB_ACCOUNTS - i));
         }
 
         assertEq(heap.size, NB_ACCOUNTS / 2);
@@ -211,11 +211,11 @@ contract TestHeapOrdering is DSTest {
 
     function testShouldRemoveAllSortedAccount() public {
         for (uint256 i = 0; i < accounts.length; i++) {
-            update(accounts[i], 0, NB_ACCOUNTS - i);
+            update(accounts[i], 0, uint96(NB_ACCOUNTS - i));
         }
 
         for (uint256 i = 0; i < accounts.length; i++) {
-            update(accounts[i], NB_ACCOUNTS - i, 0);
+            update(accounts[i], uint96(NB_ACCOUNTS - i), 0);
         }
 
         assertEq(heap.getHead(), ADDR_ZERO);
@@ -223,10 +223,10 @@ contract TestHeapOrdering is DSTest {
     }
 
     function testShouldInsertAccountSortedAtTheBeginningUntilNDS() public {
-        uint256 value = 50;
+        uint96 value = 50;
 
         // Add first 10 accounts with decreasing value.
-        for (uint256 i = 0; i < 10; i++) {
+        for (uint96 i = 0; i < 10; i++) {
             update(accounts[i], 0, value - i);
         }
 
@@ -283,7 +283,7 @@ contract TestHeapOrdering is DSTest {
     function testComputeSizeBig() public {
         MAX_SORTED_USERS = 30;
 
-        for (uint256 i = 0; i < NB_ACCOUNTS; i++) {
+        for (uint96 i = 0; i < NB_ACCOUNTS; i++) {
             update(accounts[i], 0, i + 1);
         }
         // Test that the size has been increased to MAX_SORTED_USERS/2 and increased again.
@@ -401,16 +401,16 @@ contract TestHeapOrdering is DSTest {
     }
 
     function testInsertLast() public {
-        for (uint256 i; i < 10; i++) update(accounts[i], 0, NB_ACCOUNTS - i);
+        for (uint256 i; i < 10; i++) update(accounts[i], 0, uint96(NB_ACCOUNTS - i));
 
-        for (uint256 i = 10; i < 15; i++) update(accounts[i], 0, i - 9);
+        for (uint256 i = 10; i < 15; i++) update(accounts[i], 0, uint96(i - 9));
 
         for (uint256 i = 10; i < 15; i++) assertLe(heap.accounts[i].value, 10);
     }
 
     function testInsertWrap() public {
         MAX_SORTED_USERS = 20;
-        for (uint256 i = 0; i < 20; i++) update(accounts[i], 0, NB_ACCOUNTS - i);
+        for (uint256 i = 0; i < 20; i++) update(accounts[i], 0, uint96(NB_ACCOUNTS - i));
 
         update(accounts[20], 0, 1);
 
@@ -419,7 +419,7 @@ contract TestHeapOrdering is DSTest {
 
     function testDecreaseRankChanges() public {
         MAX_SORTED_USERS = 4;
-        for (uint256 i = 0; i < 16; i++) update(accounts[i], 0, 20 - i);
+        for (uint256 i = 0; i < 16; i++) update(accounts[i], 0, uint96(20 - i));
 
         uint256 rank5Before = heap.ranks[accounts[5]];
         uint256 rank0Before = heap.ranks[accounts[0]];
@@ -438,7 +438,7 @@ contract TestHeapOrdering is DSTest {
     }
 
     function testIncreaseRankChange() public {
-        for (uint256 i = 0; i < 20; i++) update(accounts[i], 0, 20 - i);
+        for (uint256 i = 0; i < 20; i++) update(accounts[i], 0, uint96(20 - i));
 
         MAX_SORTED_USERS = 10;
 
@@ -450,7 +450,7 @@ contract TestHeapOrdering is DSTest {
     }
 
     function testIncreaseRankChangeShiftUp() public {
-        for (uint256 i = 0; i < 20; i++) update(accounts[i], 0, 20 - i);
+        for (uint256 i = 0; i < 20; i++) update(accounts[i], 0, uint96(20 - i));
 
         MAX_SORTED_USERS = 10;
 
@@ -477,9 +477,9 @@ contract TestHeapOrdering is DSTest {
     }
 
     function testRemoveShiftDown() public {
-        for (uint256 i = 0; i < 20; i++) update(accounts[i], 0, NB_ACCOUNTS - i);
+        for (uint256 i = 0; i < 20; i++) update(accounts[i], 0, uint96(NB_ACCOUNTS - i));
 
-        update(accounts[5], NB_ACCOUNTS - 5, 0);
+        update(accounts[5], uint96(NB_ACCOUNTS - 5), 0);
 
         assertEq(heap.accounts[5].value, NB_ACCOUNTS - 2 * (5 + 1) + 1);
     }
